implement log loading

+++ src/Packets.h
@@ -5,7 +5,7 @@
 #include <stdbool.h>
 #define PACKETS(P, N)                                                          \
   P(N, enabled, bool)                                                          \
-  P(N, state, uint16_t)                                                        \
+  P(N, state, uint32_t)                                                        \
   P(N, controlmode, uint16_t)                                                  \
   P(N, time, double)
